myapp/requirements.txt
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Deprecated==1.2.13
docutils==0.19
firebase-rest-api==1.10.1
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
idna==3.4
jwcrypto==1.4.2
Kivy==2.1.0
Kivy-Garden==0.1.5
kivy-garden.xcamera==2020.613
kivy-garden.zbarcam==2020.524
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.5.0
pkce==1.0.3
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
python-jwt==4.0.0
pyzbar==0.1.9
requests==2.28.2
rsa==4.9
six==1.16.0
urllib3==1.26.15
wrapt==1.15.0
xcamera==2020.613
zbarcam==2020.524
